requirements/common.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Pillow==7.2.0
PyMySQL==0.10.0
python-dotenv==0.13.0
gunicorn==20.0.4

redis==3.5.3
celery==4.4.7
cryptography

git+https://gitlab.studyxiao.cn/studyxiao/haw.git#egg=haw
